requirements.txt
python-dotenv
# confluence
atlassian-python-api
beautifulsoup4
PyMuPDF

ray
torch
transformers
langchain
langchain-community
langchain-huggingface
sentence-transformers
sentencepiece

# Inmemory Vector DB
faiss-cpu

# PostgreSQL Vector DB
asyncpg
pgvector
psycopg[binary,pool]
psycopg2-binary
sqlalchemy[asyncio]
